Make React a peer dependency (#71)
This PR is in response to this support thread, where, because we have an internal dependency on React, it requires our customers to bundle two versions of React.

There's a peerDependencies field that we can use in our package.json that tells the package to use the version of React in the implementing app, as long as it satisfies the version constraints (in this case, React 16.8+ or React 17.0+). This should prevent the installation of multiple Reacts.